requirements.txt
# --- Core Bot Libraries ---
nextcord==2.3.2
python-dotenv==1.0.0
aiosqlite==0.19.0

# --- Web Dashboard Libraries ---
Flask==2.3.3
Flask-SQLAlchemy==3.0.5

# --- Other Utilities ---
apscheduler==3.10.4
requests==2.31.0